requirements.txt
# ShopVista Multi-Language Customer Support Automation
# Requires: Python 3.10+

# Embedding & Vector Database
sentence-transformers>=2.2.0
chromadb>=0.4.0

# RAG Pipeline & LLM
google-genai>=1.0.0
python-dotenv>=1.0.0
langdetect>=1.0.9
